restructure code and handle chango of storage dir
